third_party/WebKit/LayoutTests/http/tests/locks/clientids-expected.txt

CONSOLE ERROR: line 13: Uncaught Error: Reloading with appropriate origin for desired context...
This is a testharness.js-based test.
FAIL Client IDs match between Locks API and Service Workers assert_true: clientIds should match, but are different expected true got false
Harness: the test ran to completion.
